Upstream merge B5a: auto-merged 54 differs files (3-way merge, no conflicts)
For each file where linux had local edits AND upstream Release 37 also
modified it, ran git merge-file (3-way merge: local + upstream-pre +
upstream-post). 54 of the 74 differs merged cleanly with no conflict
markers — these are the cases where linux's edits and upstream's
edits touched disjoint regions.

20 files still have conflict markers and will be resolved in
follow-up commits (B5b-B5e), grouped by category:
- .rc / main.cpp / version_config.hpp pattern conflicts
- Color.qss / Font.qss
- GUI consolidation overlaps
- Security files (openssl-deprecation territory)
- CMakeLists.txt (deferred to B7)

Diff-algorithm: patience.

+++ libraries/opensyde_gui/src/system_definition/node_edit/data_blocks/C_SdNdeDbSelectDataPools.cpp
@@ -193,7 +193,7 @@ void C_SdNdeDbSelectDataPools::m_OnSearch(const QString & orc_Text) const
    if (orc_Text.isEmpty() == true)
    {
       m_HandleHiding();
-      m_UpdateSelection(this->mpc_Ui->pc_TreeView->GetSelectedDataElements().size());
+      m_UpdateSelection(static_cast<int32_t>(this->mpc_Ui->pc_TreeView->GetSelectedDataElements().size()));
    }
    else
    {
@@ -204,7 +204,7 @@ void C_SdNdeDbSelectDataPools::m_OnSearch(const QString & orc_Text) const
       }
       else
       {
-         m_UpdateSelection(this->mpc_Ui->pc_TreeView->GetSelectedDataElements().size());
+         m_UpdateSelection(static_cast<int32_t>(this->mpc_Ui->pc_TreeView->GetSelectedDataElements().size()));
       }
       this->mpc_Ui->pc_GroupBoxInitialDataElementNoElements->setVisible(false);
    }
